Setting up DDL table structure as well as initial Scaffold Commands.
Created the scaffold and migrate commands to set up the DDL
structure for the user and wine table as well as establish the one to
many relationship between a user and a wine. In addition, set up the
mandatory and non mandatory fields in the wines table.
Also created a second script to establish the 1:N relationship between
user and wine.